// src/ArquivoMate2.Ui/src/app/client/models/document-feature-processing-dto-api-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DocumentFeatureProcessingDto } from '../models/document-feature-processing-dto';
export interface DocumentFeatureProcessingDtoApiResponse {
  data?: DocumentFeatureProcessingDto;
  errorCode?: string | null;
  errors?: ({
[key: string]: Array<string>;
}) | null;
  message?: string | null;
  success?: boolean;
  timestamp?: string;
}